pyproject.toml
[project]
name = "context-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.165.1",
    "crewai-tools>=0.0.1",
    "ipykernel>=6.30.1",
    "groq>=0.15.0",
    "chromadb>=1.1.1",
    "pymupdf>=1.26.0",
    "voyageai>=0.3.4",
    "zep-crewai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "firecrawl-py>=4.3.0",
    "pyyaml>=6.0.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
